chore: extract curve to separate package
